kernel: Boost CPU to the max when launcher3 becomes a top app
Android moves the launcher3 process to the top-app cpuset when the
multitasking view is entered since this launcher handles the multitasking
interface, which is a very expensive interface on Pixel devices. Because of
this, when rapidly hitting the multitask button and switching apps
sometimes, there's noticeably bad stuttering due to the overhead from the
app-switch interface itself.

In order to fix the stuttering when multitasking, boost all CPUs and the
DRAM bus to the max when the Pixel launcher is moved to the top-app cpuset.

Change-Id: I2adab7128b59e5c99fde250bbd95321bcc076821
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>